Fix resource optimization bug (#1317)
* fix resource optimization bug

* address commments

+++ kubernetes/linux/main.sh
@@ -1097,29 +1097,28 @@ else
       setGlobalEnvVar AZMON_RESOURCE_OPTIMIZATION_ENABLED "${AZMON_RESOURCE_OPTIMIZATION_ENABLED}"
 fi
 
-if [ "$AZMON_RESOURCE_OPTIMIZATION_ENABLED" != "true" ]; then
-      # no dependency on fluentd for prometheus side car container
-      if [ "${CONTAINER_TYPE}" != "PrometheusSidecar" ] && [ "${GENEVA_LOGS_INTEGRATION_SERVICE_MODE}" != "true" ]; then
-            if [ ! -e "/etc/config/kube.conf" ]; then
-                  if [ "$LOGS_AND_EVENTS_ONLY" != "true" ]; then
-                        echo "*** starting fluentd v1 in daemonset"
-                        if [ "${ENABLE_CUSTOM_METRICS}" == "true" ]; then
-                              mv /etc/fluent/container-cm.conf /etc/fluent/container.conf
-                        fi
-                        fluentd -c /etc/fluent/container.conf -o /var/opt/microsoft/docker-cimprov/log/fluentd.log --log-rotate-age 5 --log-rotate-size 20971520 &
-                  else
-                        echo "Skipping fluentd since LOGS_AND_EVENTS_ONLY is set to true"
-                  fi
-            else
-                  echo "*** starting fluentd v1 in replicaset"
-                  if [ "${ENABLE_CUSTOM_METRICS}" == "true" ]; then
-                        mv /etc/fluent/kube-cm.conf /etc/fluent/kube.conf
-                  fi
-                  fluentd -c /etc/fluent/kube.conf -o /var/opt/microsoft/docker-cimprov/log/fluentd.log --log-rotate-age 5 --log-rotate-size 20971520 &
+#start fluentd
+if [ "${CONTROLLER_TYPE}" == "ReplicaSet" ] && [ "${GENEVA_LOGS_INTEGRATION_SERVICE_MODE}" != "true" ]; then
+    echo "*** starting fluentd v1 in replicaset"
+    if [ "${ENABLE_CUSTOM_METRICS}" == "true" ]; then
+        mv /etc/fluent/kube-cm.conf /etc/fluent/kube.conf
+    fi
+    fluentd -c /etc/fluent/kube.conf -o /var/opt/microsoft/docker-cimprov/log/fluentd.log --log-rotate-age 5 --log-rotate-size 20971520 &
+elif [ "$AZMON_RESOURCE_OPTIMIZATION_ENABLED" != "true" ]; then
+    # no dependency on fluentd for Prometheus sidecar container
+    if [ "${CONTAINER_TYPE}" != "PrometheusSidecar" ] && [ "${GENEVA_LOGS_INTEGRATION_SERVICE_MODE}" != "true" ] && [ ! -e "/etc/config/kube.conf" ]; then
+        if [ "$LOGS_AND_EVENTS_ONLY" != "true" ]; then
+            echo "*** starting fluentd v1 in daemonset"
+            if [ "${ENABLE_CUSTOM_METRICS}" == "true" ]; then
+                mv /etc/fluent/container-cm.conf /etc/fluent/container.conf
             fi
-      fi
+            fluentd -c /etc/fluent/container.conf -o /var/opt/microsoft/docker-cimprov/log/fluentd.log --log-rotate-age 5 --log-rotate-size 20971520 &
+        else
+            echo "Skipping fluentd since LOGS_AND_EVENTS_ONLY is set to true"
+        fi
+    fi
 else
-      echo "Skipping fluentd since AZMON_RESOURCE_OPTIMIZATION_ENABLED is set to ${AZMON_RESOURCE_OPTIMIZATION_ENABLED}"
+    echo "Skipping fluentd for linux daemonset since AZMON_RESOURCE_OPTIMIZATION_ENABLED is set to ${AZMON_RESOURCE_OPTIMIZATION_ENABLED}"
 fi
 
 #If config parsing was successful, a copy of the conf file with replaced custom settings file is created
